Create Focus.py
Need to create a filter system to log the apps that arent in the whitelist. Once this is done find a way to plug those into the script and terminate those processes. (Phase 1)

Channel your inner front end and work on a interface that allows when you would like the program to run and what apps you would like to add to the whitelist
(Phase 2)

Market
(Phase 3)